% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunc.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n = 10, k = 3)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of successes}
}
\value{
the number of combinations calculated
}
\description{
Calculates the number of combinations in which k successes occur within n trials
}
\examples{

# default
bin_choose()

# another example
bin_choose(n = 5, k = 3)

}
