% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{binomial Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability that a trial will succeed}
}
\value{
computed cumulative probability distribution
}
\description{
compute the cumulative probability distribution of a binomial distribution
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)

}
