% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{binomial distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability that a trial will succeed}
}
\value{
a data frame with the probability distribution
}
\description{
compute the probability distribution of a binomial distribution
}
\examples{
bin_distribution(trials = 5, prob = 0.5)

}
