% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{binomial Kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability that a trial will succeed}
}
\value{
computed kurtosis
}
\description{
compute the kurtosis of a binomial distribution
}
\examples{
bin_kurtosis(trials = 5, prob = 0.3)
}
