% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{binomial Mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability that a trial will succeed}
}
\value{
computed mean
}
\description{
compute the expected mean of a binomial distribution
}
\examples{
bin_mean(trials = 5, prob = 0.3)
}
