% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainfunc.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success = 2, trials = 5, prob = 0.5)
}
\arguments{
\item{success}{number of successes}

\item{trials}{number of trials}

\item{prob}{probability of success within one trial}
}
\value{
the probability of the given case
}
\description{
Calculates the probability of a given number of successes in a given numbers of trials with a specified probability
}
\examples{

# default
bin_probability()

# another example
bin_probability(success = 1, trials = 3, prob = 0.5)

}
