% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{binomial Skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability that a trial will succeed}
}
\value{
computed skewness
}
\description{
compute the skewness of a binomial distribution
}
\examples{
bin_skewness(trials = 5, prob = 0.3)
}
