% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{binomial Variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability that a trial will succeed}
}
\value{
binomial object
}
\description{
construct a binomial object with given parameters
}
\examples{
bin_variable(trials = 5, prob = 0.45)
}
