% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{check_prob}
\alias{check_prob}
\title{Checking "Prob"}
\usage{
check_prob(prob)
}
\arguments{
\item{prob}{vector of (at least 2) probability values, between 0 and 1, summing to 1}
}
\value{
logical value; TRUE or FALSE.
}
\description{
Checks if the given input is a correct "sides" argument
}
