% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{check_sides}
\alias{check_sides}
\title{Checking "Sides"}
\usage{
check_sides(sides)
}
\arguments{
\item{sides}{vector of (at least 2) sides, default of 1 and 2e}
}
\value{
logical value; TRUE or FALSE.
}
\description{
Checks if the given input is a correct "sides" argument
}
