% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{device}
\alias{device}
\title{device}
\usage{
device(sides = c(1, 2), prob = c(0.5, 0.5))
}
\arguments{
\item{sides}{vector of (at least 2) sides, default of 1 and 2}

\item{prob}{probability for each side}
}
\value{
device object
}
\description{
Creates an object of class device
}
