% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_choose.R
\name{bin_choose}
\alias{bin_choose}
\title{Bin Choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{(trials)}

\item{k}{(success)}
}
\value{
number of combinations
}
\description{
calculates the number of combinations in which k successes can occur in n trials
}
\examples{
#5 choose 3
bin_choose(n = 5, k = 2)
}
