% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial Choose Function}
\usage{
bin_choose(n, k)
}
\arguments{
\item{trials}{number of trials}

\item{success}{number of desired successes}
}
\value{
number of successes in trials
}
\description{
Find number of successes in given number of trials given probability p
}
\examples{
bin_choose(n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)

}
