% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{bin_choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{input trials number}

\item{k}{how many trials to choose}
}
\value{
the result
}
\description{
a measure of the ¡°tailedness¡± of the probability distribution of a random variable
}
\examples{
bin_choose(n = 5, k = 2)
}
