% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{Function bin_choose()}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{is the number of (fixed) trials}

\item{k}{is a variable that represents the number of success out of n trials}
}
\value{
the number of combinatons in which successes(k) and trials(n) can occur
}
\description{
using factorial to caculate the number combinations in which successes an occur in trials
}
\examples{
bin_choose (n = 5, k = 2)
bin_choose(5, 0)
bin_choose(5, 1:3)
}
