% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_choose}
\alias{bin_choose}
\title{K choose K}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{the value of n}

\item{k}{the value of k}
}
\value{
computed value of n choose k
}
\description{
compute the value of (n choose k)
}
\examples{
bin_choose(n=5, k=2)

bin_choose(5, 0)

bin_choose(5, 1:3)

}
