% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_choose.R
\name{bin_choose}
\alias{bin_choose}
\title{Binomial choose}
\usage{
bin_choose(n, k)
}
\arguments{
\item{n}{number of trials}

\item{k}{number of success}
}
\value{
the number of combinations
}
\description{
Calculates the number of combinations
}
\examples{
 \dontrun{
bin_choose(n=5L,k=2L)
bin_choose(5L,0L)
bin_choose(5L,1L:3L)
 }
}
