% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_cumulative.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Bin Cumulative}
\usage{
bin_cumulative(trials, prob)
}
\value{
probability distribution and cumulative probabilities
}
\description{
return data frame with probability distribution and cumulative probabilities
}
\examples{
# Build a binomial distribution for 5 trials, prob =0.5
dis2 <- bin_cumulative(trials=5,prob=0.5)
dis2
#plot distribution
plot(dis2)
}
