% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Binomial Cumulative Function}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{prob}{p of success}
}
\value{
df with "dataframe" and bincum as classes
}
\description{
creates binomial distributive function data frame
}
