% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{input trials number}

\item{prob}{probability}
}
\value{
a data frame with sucesses in the first column, probability in the second column, cummulative in the third
}
\description{
a dataframe which list the success, the probability and the cumulative probability
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
}
