% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{Function bin_cumulative()}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{is a non-negative integer; numeric vector}

\item{prob}{is the probaility of success on each trial; numeric vector}
}
\value{
a data frame with probability distributin and cumulative probability
}
\description{
it takes two aruments trials & prob and calculates the cumulated probability of each trial
}
\examples{
bin_cum <- bin_cumulative(trials = 5, prob = 0.5)
plot(bin_cum)
}
