% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Binomial distribution Function}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{p of success}
}
\value{
dataframe with columns ("bindis" and "data.frame)
}
\description{
makes binomial distrubiton data frame given the parameters entered
}
\examples{
bin_distribution(trials = 5, prob = 0.5)

}
