% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{bin_distribution}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{input trials number}

\item{prob}{probability}
}
\value{
a data frame with the probability distribution: sucesses in the first column, probability in the second column.
}
\description{
a dataframe which list the probability of given parameters
}
\examples{
bin_distribution(trials = 5, prob = 0.5)
}
