% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_distribution}
\alias{bin_distribution}
\title{Function bin_distribution()}
\usage{
bin_distribution(trials, prob)
}
\arguments{
\item{trials}{is a non-negative integer; numeric vector}

\item{prob}{is the probaility of success on each trial;numeric vector}
}
\value{
a data frame with probability distribution
}
\description{
it takes two aruments trials & prob to calculate the probability distribution of success in a given trial and given probability.
}
\examples{
bin_dis <- bin_distribution(trials = 5, prob = 0.5)
plot(bin_dis)
}
