% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{p of success on any given trial}
}
\value{
the kurtosis
}
\description{
calculuates the kurtosis of a binomial distriubtion
}
\examples{
bin_kurtosis(10, 0.3)
}
