% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the kurtosis of the input
}
\description{
check if the input is valid and compute the kurtosis of the input using aux_kurtosis
}
\examples{
bin_kurtosis(10, 0.3)
}
