% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{the probability of success every time}
}
\value{
the mean given the trials and probability
}
\description{
calc mean of binomial distribution
}
\examples{
bin_mean(10, 0.3)
}
