% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the mean of the input
}
\description{
check if the input is valid and compute the mean of the input using aux_mean
}
\examples{
bin_mean(10, 0.3)
}
