% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_mean.R
\name{bin_mean}
\alias{bin_mean}
\title{Binomial mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
the mean of binomial
}
\description{
Calculates the mean of binomial
}
\examples{
 \dontrun{
bin_mean(trials=10L,prob=0.3)
 }
}
