% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_choose}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{how many trials to choose}

\item{trials}{input trials number}

\item{prob}{probability}
}
\value{
the result
}
\description{
a measure of the ¡°tailedness¡± of the probability distribution of a random variable
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
}
