% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_probability}
\alias{bin_probability}
\title{Function bin_probability()}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{is a non-negative interger less than of equal to trials and check_sucess function is used to check if success is valid}

\item{trials}{is a non-negative integer and check_trial function is used to check if triasl is valid}

\item{prob}{is the probaility of success on each trial and check_prof is used to check if pro is valid}
}
\value{
binomial probability of successes in trials given by probability
}
\description{
it takes three arguemnts sucess, trials, and prob to calculate bionomial probability
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(success = 0:2, trials = 5, prob = 0.5)
bin_probability(success = 55, trials = 100, prob = 0.45)
}
