% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{p of success on any given trial}
}
\value{
the skewness
}
\description{
calculuates the skewness of a binomial distriubtion
}
\examples{
bin_skewness(10, 0.3)
}
