% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability of success}
}
\value{
the skewness of the input
}
\description{
check if the input is valid and compute the skewness of the input using aux_skewness
}
\examples{
bin_skewness(10, 0.3)
}
