% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{Bin Variable}
\usage{
bin_variable(trials, prob)
}
\value{
list with number of trials and probability of success
}
\description{
returns binomial random variable object
}
\examples{
#Build a binomial distribution
bin <- bin_variable(10,0.3)
# summarize
binsum <- summary(bin)
# print the summary
binsum
}
