% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial CVariable Function}
\usage{
bin_variable(prob, trials)
}
\arguments{
\item{prob}{probability}

\item{trials}{trials}
}
\value{
binomial variable object that stores trials and probability
}
\description{
creates binomial variable object that stores trials and probability
}
