% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variable}
\alias{bin_variable}
\title{Function bin_variable()}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{is a non-negative integer; numeric vector}

\item{prob}{is the probaility of success on each trial; numeric vector}
}
\value{
a data frame with a list of trials and probablity
}
\description{
it takes two aruments trials & prob and calculates an object of class caled "binvar" which is a binomial random variable object
}
\examples{
bin_var <- bin_variable(trials = 5, prob = 0.5)
bin_var

summary(bin_var)
}
