% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variable.R
\name{bin_variable}
\alias{bin_variable}
\title{Binomial variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
the binomial variable
}
\description{
Calculates the binomial variable
}
\examples{
 \dontrun{
bin_variable(trials=10,prob=0.3)
 }
}
