% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{total number of trials}

\item{prob}{prob of success for any given trial}
}
\value{
the variance
}
\description{
calc variance of binomial distribution
}
\examples{
bin_variance(10, 0.3)
}
