% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_variance.R
\name{bin_variance}
\alias{bin_variance}
\title{Binomial variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of trials}

\item{prob}{probability}
}
\value{
the variance of binomial
}
\description{
Calculates the variance of binomial
}
\examples{
 \dontrun{
bin_variance(trials=10L,prob=0.3)
 }
}
